requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
requests
google-genai
